requirements.txt
APScheduler==3.10.4
django-apscheduler
asgiref==3.8.1
attrs==24.2.0
boto3==1.26.66
botocore==1.29.165
cfgv==3.4.0
distlib==0.3.9
Django==4.2.6
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.23.1
filelock==3.16.1
identify==2.6.1
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nodeenv==1.9.1
pillow
platformdirs==4.3.6
pre-commit==2.20.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.5.1
toml==0.10.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
virtualenv==20.27.1
whitenoise==6.0.0
django-cors-headers==4.5.0
langchain
langchain-openai
langchain-core
python-dotenv
gunicorn==21.2.0

# drf-writable-nested==0.7.1
# language_tool_python==2.9.0
# java too for the spellchecker openjdk-17-jdk
